Change contraint namespace and move to constraint folder, add null check in validator, fix code style bugs
