+++ src/Validator/FileMatchesTypeValidator.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace BitBag\SyliusCmsPlugin\Validator;
 
 use BitBag\SyliusCmsPlugin\Entity\Media;
+use BitBag\SyliusCmsPlugin\Validator\Constraint\FileMatchesType;
 use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\ConstraintValidator;
 use Symfony\Component\Validator\Exception\UnexpectedTypeException;
@@ -17,7 +18,11 @@ final class FileMatchesTypeValidator extends ConstraintValidator
             throw new UnexpectedTypeException($constraint, FileMatchesType::class);
         }
 
-        $mime = $value->getFile() ? $value->getFile()->getMimeType() : $value->getMimeType();
+        if ($value->hasFile() && empty($value->getFile()->getMimeType())) {
+            return;
+        }
+
+        $mime = $value->hasFile() ? $value->getFile()->getMimeType() : $value->getMimeType();
 
         if ($value->getType() === Media::IMAGE_TYPE && !(str_starts_with($mime, 'image/'))) {
             $this->context->buildViolation($constraint->messageImage)
